ReturnItem as interface #1053
